+++ packages/dialog/src/react/__stories__/index.story.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { action } from '@storybook/addon-actions'
-import { storiesOf } from '@storybook/react'
+import { Meta, Story } from '@storybook/react/types-6-0'
 
 import { Button } from '@pluralsight/ps-design-system-button'
 import * as Text from '@pluralsight/ps-design-system-text'
@@ -44,21 +44,34 @@ const MockDialog: React.FC<React.ComponentProps<typeof Dialog>> = props => {
   return <Dialog {...props}>{children}</Dialog>
 }
 
-storiesOf('onClose', module)
-  .add('with onClose', () => <MockDialog onClose={closeAction} />)
-  .add('with disableCloseButton', () => (
-    <MockDialog disableCloseButton>
-      <Text.Heading>
-        <h1>Wowzers, a Dialog</h1>
-      </Text.Heading>
-    </MockDialog>
-  ))
+export default {
+  title: 'Components/Dialog',
+  component: Dialog
+} as Meta
 
-const positionStories = storiesOf('tailPosition', module)
-positionStories.add('none', () => <MockDialog />)
+export const OnClose: Story = () => <MockDialog onClose={closeAction} />
 
-Object.values(Dialog.tailPositions).forEach(pos =>
-  positionStories.add(pos, () => <MockDialog tailPosition={pos} />)
+export const OnCloseDisableClosebutton: Story = () => (
+  <MockDialog disableCloseButton>
+    <Text.Heading>
+      <h1>Wowzers, a Dialog</h1>
+    </Text.Heading>
+  </MockDialog>
+)
+
+export const TailPositions: Story = () => (
+  <div style={{ maxWidth: '50%', margin: '0 auto' }}>
+    <StoryGrid>
+      <div>none</div>
+      <MockDialog />
+      {Object.values(Dialog.tailPositions).map(position => (
+        <React.Fragment key={position}>
+          <div>{position}</div>
+          <MockDialog tailPosition={position} />
+        </React.Fragment>
+      ))}
+    </StoryGrid>
+  </div>
 )
 
 interface ModalStoryProps extends HTMLPropsFor<'div'> {
@@ -117,79 +130,85 @@ const ModalStory: React.FC<ModalStoryProps> = ({ children }) => {
   )
 }
 
-storiesOf('modal', module)
-  .add('default', () => (
-    <ModalStory>{props => <MockDialog {...props} />}</ModalStory>
-  ))
-  .add('no close button', () => (
-    <ModalStory>
-      {props => <MockDialog {...props} disableCloseButton />}
-    </ModalStory>
-  ))
-  .add('no focus on mount', () => (
-    <ModalStory>
-      {props => <MockDialog {...props} disableFocusOnMount />}
-    </ModalStory>
-  ))
-  .add('no click overlay', () => (
-    <ModalStory>
-      {props => <MockDialog {...props} disableCloseOnOverlayClick />}
-    </ModalStory>
-  ))
-  .add('no escape key', () => (
-    <ModalStory>
-      {props => <MockDialog {...props} disableCloseOnEscape />}
-    </ModalStory>
-  ))
-  .add('no focusable child elements', () => (
-    <ModalStory>
-      {props => (
-        <MockDialog {...props}>
-          <Text.Heading>
-            <h1>Wowzers, a Dialog</h1>
-          </Text.Heading>
-        </MockDialog>
-      )}
-    </ModalStory>
-  ))
-  .add('overflow-y', () => (
-    <ModalStory>
-      {props => (
-        <MockDialog {...props}>
-          {Array.from(Array(20)).map((_, i) => (
-            <Text.P key={i}>
-              <a href="#">Brownie bear claw</a> liquorice dragée candy canes
-              pastry topping. Chocolate cake soufflé sweet roll jelly beans oat
-              cake donut. Wafer chocolate cake pastry chocolate bar fruitcake.
-              Cupcake jelly-o croissant lollipop liquorice.{' '}
-              <a href="#">Tart donut</a> lollipop dragée tootsie roll wafer
-              lemon drops cupcake chocolate bar.
-            </Text.P>
-          ))}
-        </MockDialog>
-      )}
-    </ModalStory>
-  ))
-  .add('overflow-y with tail', () => (
-    <ModalStory>
-      {props => (
-        <MockDialog {...props} tailPosition={Dialog.tailPositions.topCenter}>
-          {Array.from(Array(20)).map((_, i) => (
-            <Text.P key={i}>
-              <a href="#">Brownie bear claw</a> liquorice dragée candy canes
-              pastry topping. Chocolate cake soufflé sweet roll jelly beans oat
-              cake donut. Wafer chocolate cake pastry chocolate bar fruitcake.
-              Cupcake jelly-o croissant lollipop liquorice.{' '}
-              <a href="#">Tart donut</a> lollipop dragée tootsie roll wafer
-              lemon drops cupcake chocolate bar.
-            </Text.P>
-          ))}
-        </MockDialog>
-      )}
-    </ModalStory>
-  ))
+export const ModalDefault: Story = () => (
+  <ModalStory>{props => <MockDialog {...props} />}</ModalStory>
+)
 
-storiesOf('stylesFor', module).add('dialog_content', () => (
+export const ModalNoClosebutton: Story = () => (
+  <ModalStory>
+    {props => <MockDialog {...props} disableCloseButton />}
+  </ModalStory>
+)
+
+export const ModalNoFocusOnMount: Story = () => (
+  <ModalStory>
+    {props => <MockDialog {...props} disableFocusOnMount />}
+  </ModalStory>
+)
+
+export const ModalNoClickOverlay: Story = () => (
+  <ModalStory>
+    {props => <MockDialog {...props} disableCloseOnOverlayClick />}
+  </ModalStory>
+)
+
+export const ModalNoEscapeKey: Story = () => (
+  <ModalStory>
+    {props => <MockDialog {...props} disableCloseOnEscape />}
+  </ModalStory>
+)
+
+export const ModalNoFocusableChildElements: Story = () => (
+  <ModalStory>
+    {props => (
+      <MockDialog {...props}>
+        <Text.Heading>
+          <h1>Wowzers, a Dialog</h1>
+        </Text.Heading>
+      </MockDialog>
+    )}
+  </ModalStory>
+)
+
+export const ModalOverlowY: Story = () => (
+  <ModalStory>
+    {props => (
+      <MockDialog {...props}>
+        {Array.from(Array(20)).map((_, i) => (
+          <Text.P key={i}>
+            <a href="#">Brownie bear claw</a> liquorice dragée candy canes
+            pastry topping. Chocolate cake soufflé sweet roll jelly beans oat
+            cake donut. Wafer chocolate cake pastry chocolate bar fruitcake.
+            Cupcake jelly-o croissant lollipop liquorice.{' '}
+            <a href="#">Tart donut</a> lollipop dragée tootsie roll wafer lemon
+            drops cupcake chocolate bar.
+          </Text.P>
+        ))}
+      </MockDialog>
+    )}
+  </ModalStory>
+)
+
+export const ModalOverflowYWithTail: Story = () => (
+  <ModalStory>
+    {props => (
+      <MockDialog {...props} tailPosition={Dialog.tailPositions.topCenter}>
+        {Array.from(Array(20)).map((_, i) => (
+          <Text.P key={i}>
+            <a href="#">Brownie bear claw</a> liquorice dragée candy canes
+            pastry topping. Chocolate cake soufflé sweet roll jelly beans oat
+            cake donut. Wafer chocolate cake pastry chocolate bar fruitcake.
+            Cupcake jelly-o croissant lollipop liquorice.{' '}
+            <a href="#">Tart donut</a> lollipop dragée tootsie roll wafer lemon
+            drops cupcake chocolate bar.
+          </Text.P>
+        ))}
+      </MockDialog>
+    )}
+  </ModalStory>
+)
+
+export const ModalStylesFor: Story = () => (
   <ModalStory>
     {props => (
       <MockDialog
@@ -203,4 +222,20 @@ storiesOf('stylesFor', module).add('dialog_content', () => (
       />
     )}
   </ModalStory>
-))
+)
+
+const StoryGrid: React.FC<{ cols?: number }> = props => {
+  const { cols = 2, ...rest } = props
+
+  return (
+    <div
+      style={{
+        display: 'grid',
+        gap: '20px',
+        gridTemplateColumns: Array(cols).fill('1fr').join(' '),
+        justifyItems: 'left'
+      }}
+      {...rest}
+    />
+  )
+}