+++ packages/layout/src/react/__specs__/page-width-layout.spec.tsx
@@ -1,0 +1,29 @@
+import { convertStoriesToJestCases } from '@pluralsight/ps-design-system-util'
+import { render } from '@testing-library/react'
+import { axe } from 'jest-axe'
+import React from 'react'
+
+import { PageWidthLayout } from '../index'
+import * as stories from '../__stories__/page-width-layout.story'
+
+describe('PageWidthLayout', () => {
+  const cases = convertStoriesToJestCases(stories)
+
+  it('renders', () => {
+    const { getByTestId } = render(
+      <PageWidthLayout data-testid="undertest">
+        <div>something special</div>
+      </PageWidthLayout>
+    )
+
+    expect(getByTestId('undertest')).toBeInTheDocument()
+  })
+
+  describe.each(cases)('%s story', (_name, Story) => {
+    it('has no axe-core violations', async () => {
+      const { container } = render(<Story {...Story.args} />)
+      const results = await axe(container)
+      expect(results).toHaveNoViolations()
+    })
+  })
+})