+++ packages/row/src/react/__specs__/index.spec.tsx
@@ -1,0 +1,33 @@
+import { convertStoriesToJestCases } from '@pluralsight/ps-design-system-util'
+import { screen, render } from '@testing-library/react'
+import { axe } from 'jest-axe'
+import React from 'react'
+
+import { Row } from '../index'
+import * as stories from '../__stories__/index.story'
+
+describe('Row', () => {
+  const cases = convertStoriesToJestCases(stories)
+
+  it('renders', () => {
+    const { getByTestId } = render(<Row data-testid="undertest" />)
+
+    expect(getByTestId('undertest')).toBeInTheDocument()
+  })
+
+  it('forwards ref', () => {
+    const ref = React.createRef<HTMLDivElement>()
+    render(<Row ref={ref}></Row>)
+    expect(ref).not.toBeNull()
+  })
+
+  describe('accessibility', () => {
+    describe.each(cases)('%s story', (_name, Story) => {
+      it('has no axe-core violations', async () => {
+        const { container } = render(<Story {...Story.args} />)
+        const results = await axe(container)
+        expect(results).toHaveNoViolations()
+      })
+    })
+  })
+})